KVM: TDX: Enumerate number of L2 VMs
TD partitioning is a feature defined since TDX1.4, which allows one TD
creating several L2 VMs. The maximum number of L2 VM is 3. This is
exposed to the userspace VMM as part of tdx_capabilities so that the
user space VMM can decide how many L2 VMs will be supported by a L1 TD.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>