KVM: TDX: Add interface to handle L2 SEPT
The L2 SEPT needs to be built by host VMM. When TDX module walks the L2
SEPT, it will exit to host VMM with EPT violation vmexit reason if some
level of L2 SEPT paging pages are missed. The extend exit qualification
will contain the vm_index and level information, which can be used by
host VMM to know which L2 SEPT needs to be built and which level of
paging is missed.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>