KVM: TDX: Add demote support for L2 SEPT
When there is L2 VM in a TD, TDH.MEM.PAGE.DEMOTE will also splite each
of the L2 SEPT if it has alias mapping with the flag ADD_ON_ALIAS == 1.

In this case, host VMM needs to allocate one L2 SEPT paging page for
each possible L2 VMs to consume. For the paging page which is not
consumed by the TDH.MEM.PAGE.DEMOTE, host VMM can free this page in the
end.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>